requirements.txt
pygame-ce
torch